Rename : modify components folder Structure
