+++ components/common/footer/index.tsx
