+++ components/main/content/hashTag/style.ts
