+++ components/main/content/title/style.ts
