+++ pages/_app.tsx
@@ -1,5 +1,5 @@
 import { AppProps } from "next/app";
-import Layout from "../components/layout";
+import Layout from "../components/common/layout";
 import { RecoilRoot } from "recoil";
 import { MDXProvider } from "@mdx-js/react";
 import CodeBlock from "../components/post/codeBlock";
